requirements.txt
# Core dependencies for working components
google-api-python-client>=2.0.0
python-dotenv>=1.0.0

# AI Framework dependencies
langchain>=0.1.0
langchain-community>=0.0.20
langchain-ollama>=0.1.0
crewai>=0.28.0
